+++ crates/pkgcraft-tools/src/bin/pk/subcmds/repo.rs
@@ -2,6 +2,7 @@ use std::process::ExitCode;
 
 use pkgcraft::config::Config;
 
+mod eapi;
 mod metadata;
 
 #[derive(Debug, clap::Args)]
@@ -18,6 +19,8 @@ impl Command {
 
 #[derive(Debug, clap::Subcommand)]
 pub enum Subcommand {
+    /// Output repo EAPI usage rates
+    Eapi(eapi::Command),
     /// Generate repo metadata
     Metadata(metadata::Command),
 }
@@ -26,6 +29,7 @@ impl Subcommand {
     fn run(&self, config: &Config) -> anyhow::Result<ExitCode> {
         use Subcommand::*;
         match self {
+            Eapi(cmd) => cmd.run(config),
             Metadata(cmd) => cmd.run(config),
         }
     }